refactor: ShortestPathEdge 위치를 도메인 내부로 이동
