+++ src/main/java/wooteco/subway/service/PathService.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Service;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.fare.FareCalculator;
-import wooteco.subway.domain.Path;
+import wooteco.subway.domain.path.Path;
 import wooteco.subway.domain.Section;
-import wooteco.subway.utils.ShortestPathEdge;
+import wooteco.subway.domain.path.ShortestPathEdge;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.PathRequest;
 import wooteco.subway.dto.response.PathResponse;
